refactor: extract YamlEditor logic into useYamlEditorController hook
